+++ wrapper/main.js
@@ -12,23 +12,32 @@ stuff
 const express = require("express");
 const app = express();
 const reqBody = require("body-parser");
+const fileUpload = require("express-fileupload");
 const morgan = require("morgan");
 const stuff = require("./static/info");
+
 
 app.use(reqBody.json());
 app.use(reqBody.urlencoded({ extended: true }));
 // HTTP logging
 app.use(morgan("dev"));
 app.use(require("./asset/list"));
+// file uploads
+app.use(fileUpload());
+
 // character
 app.use(require("./character/load"));
 app.use(require("./character/premade"));
 app.use(require("./character/redirect"));
 app.use(require("./character/save"));
 app.use(require("./character/thumb"));
+// movie
 app.use(require("./movie/list"));
 app.use(require("./movie/load"));
+app.use(require("./movie/save"));
 app.use(require("./movie/thumb"));
+app.use(require("./movie/upload"));
+// static
 app.use(require("./static/page"));
 // theme
 app.use(require("./theme/list"));
@@ -45,7 +54,7 @@ app.all("*", (req, res) => {
 			const link = (t.link || req.url);
 			const headers = t.headers;
 			for (var headerName in headers || {}) // set headers
-				res.setHeader(headerName, headers[headerName]);
+				res.set(headerName, headers[headerName]);
 			res.statusCode = t.statusCode || 200;
 			if (t.content !== undefined) res.end(t.content);
 		}